% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accuracy_stringent}
\alias{accuracy_stringent}
\title{Sequencing specificity and sensitivity}
\format{A data frame with 6 rows and 7 variables:
\describe{
  \item{exp.freq}{The expected frequency from dilution series.}
  \item{freq}{The observed frequency - set to the expected here.}
  \item{TP}{The number of true positives found out of twenty.}
  \item{sensitivity}{Sensitivity to detect mutations at this frequency}
  \item{FP}{The number of false positives in the sample}
  \item{gc_ul}{The genome copies/ul }
  \item{PPV}{Positive predictive value}
}}
\source{
data-raw/accuracy_stringent.R
}
\usage{
accuracy_stringent
}
\description{
The sensitivity and specificity of the sequencing approach applied here -(frequency cut of 2%)
determined emphirically from seqeuncing populations of 20  mutants known a priori.
}
\keyword{datasets}
