% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{betabin_sim}
\alias{betabin_sim}
\title{Simulate transmission Beta binomial model}
\usage{
betabin_sim(data, lambda, threshold, acc = accuracy_stringent)
}
\arguments{
\item{data}{a data frame with with chr,pos,freq1, and  pair_id columns}

\item{lambda}{The lambda of the zero truncated Poisson}

\item{threshold}{limit of variant calling detection}

\item{acc}{a data frame with accuracy metrics}
}
\value{
data frame the same as data but with a simulated found column
}
\description{
Simulate the transmission of alleles using the
betabinomial model. Essentially this takes
the frequency of the variant allele, and bottleneck
size and simulates the found column. It also selects a bottleneck size
based on lambda and a zero truncated Poisson. If simulation the whole data
set each pair should be run separately as to not use the same bottleneck
}
\examples{
betabin_sim(small_trans,1.2,0.02,accuracy_stringent)
}
