% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_snv.R
\name{correct_id}
\alias{correct_id}
\title{Correct numerical Ids
In our analysis the Id column refers to the sequenced id. In many cases this was
a number; but later it matched the SPECID + and extention. Therefore the column
should be read as string. However, Pandas for some reason converted the numbers to
decimal format. This function corrects that and returns a string number with out the ".0"}
\usage{
correct_id(df, col)
}
\arguments{
\item{df}{data frame or tibble to correct}

\item{col}{column that needs to be corrected.}
}
\description{
Correct numerical Ids
In our analysis the Id column refers to the sequenced id. In many cases this was
a number; but later it matched the SPECID + and extention. Therefore the column
should be read as string. However, Pandas for some reason converted the numbers to
decimal format. This function corrects that and returns a string number with out the ".0"
}
\examples{
cov_sample<-dplyr::tibble(Id=c("129.0","MH00000"))
correct_id(cov_sample,Id)
}
