% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{dist_prob}
\alias{dist_prob}
\title{Fit a distribution to pair_wise bottlenecks}
\usage{
dist_prob(data, weight, ddist, ...)
}
\arguments{
\item{data}{A tibble or data frame as outputed from trans_fit}

\item{weight}{A tibble or data frame providing the weighting factors for each
pair. It should have columns pair Id and weight_factor}

\item{ddist}{The pdf or pmf function for the distribution to test.}

\item{...}{The parameters to pass to the ddist function}
}
\value{
the negative Log likelihood weighted by the weighted data frame
}
\description{
This funciton returns the negative log likelihood for
a distribution of bottlenecks given a data frame of
fits on a per-pair basis. This can be used with mle from
the bbmle package.
}
