% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_freqs.R
\name{dist_tp}
\alias{dist_tp}
\title{Get the genetic distance between two samples .}
\usage{
dist_tp(pairs, snv)
}
\arguments{
\item{pairs}{A vector of SPECID. These strings must be found in the SPECID column of the iSNV data frame}

\item{snv}{A data frame of isnv calls from deepSNV with a SPECID column}
}
\value{
the L1 norm
}
\description{
This function compares the frequencies of alleles found in two samples and
calculates the L1 norm or manhattan distance
}
\examples{
# There  one shared iSNV here and one fixed difference the rest are the same.
print(small_isnv)
dist_tp(c("HS1595","HS1596"),small_isnv)
}
