% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_snv.R
\name{diverse_sites}
\alias{diverse_sites}
\title{Get sites with diverisity}
\usage{
diverse_sites(df, cutoff, ...)
}
\arguments{
\item{df}{variant data frame}

\item{cutoff}{integer. The allele must be found more than this many times in the
groupings below.}

\item{...}{columns to group by. There's no current need for them to be anything but
season,chr,pos,pcr_result as this allows us to}
}
\value{
the input data frame but with sites with more alleles than the cut off
}
\description{
In our analysis we are not interested in loci that only have
one allele present within a season. Cutting out such loci
helps focus on the important loci and distinctions between samples.
This function counts the unique "var" entries in a dataframe according
to the set grouping and removes sites that don't meet the cut off.
}
\examples{
diverse_sites(small_isnv,1,season,pcr_result,pos,chr)

}
