% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{dzpois}
\alias{dzpois}
\title{PMF of a zero truncated Poisson distribution}
\usage{
dzpois(x, lambda)
}
\arguments{
\item{x}{A number or vector of numbers}

\item{lambda}{The mean of the Poisson distribution that is truncated
not the mean of the truncated distribution.}
}
\value{
Logical is the number an integer
}
\description{
PMF of zero truncated Poisson distribution at a point x.
}
\examples{
dzpois(1,3)

# If Nb and l are vectors the output is a matrix of the form
# |---------Nb_j---------|
# |                      |
# |                      |
# l_i                   l_i
# |                      |
# |                      |
# |                      |
dzpois(1:2,1:3)
}
