% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_model.R
\name{g_ft}
\alias{g_ft}
\title{Likelihood of frequency starting at 0}
\usage{
g_ft(x, mu, Ne, t, gc_ul = NULL, acc = NULL)
}
\arguments{
\item{x}{The measured frequency - must be between 0 and 1}

\item{mu}{The mutation rate per nucleotide per cellular infectious cycle}

\item{Ne}{The effective population size}

\item{t}{Generations}

\item{gc_ul}{titer default to NULL use if interested in adjusting for sensitivity}

\item{acc}{sensitivity data frame default is NULL use if interested in adjusting for sensitivity}
}
\description{
This function determines the Ne and mutation rate assuming all infections begin as a clone
and then mutations are built upon this base. The probability of a site going from 0 to
x given mu, Ne, and t generations given in Rouzine 2001 JVI as
}
\details{
\deqn{g(f,t) = \frac{2 \mu N_e}{f} e^{- \frac{2N_ef}{t}}}

The output is adjusted to account for the sensitivity. This isn't important in fitting,
but it is nice in testing to know the prob of being found+ prob of not being found =1
}
\examples{
g_ft(0.5,4e-5,30,20)
}
