% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{get_close}
\alias{get_close}
\title{Finding SPECID for transmission pairs}
\usage{
get_close(meta, date, enrollid, case, iSNV = T)
}
\arguments{
\item{meta}{data frame with meta data con}

\item{date}{The estimated trasmission date}

\item{enrollid}{The enrollid}

\item{case}{c("donor","recipient")}

\item{iSNV}{boolan if TRUE then we take the sample that has iSNV if the
"best fit" doesn't. This is the historic use of the function. Recall that we
do not have samples before transmission in many cases anyway.}
}
\value{
The SPECID that best fits the critea
}
\description{
Identifies the SPECID collected nearest a given date, and on the correct side of
the date when specified. The criteria is
}
\details{
1) Sample closest to transmission that are on the "right side" of tranmission
(before transmission for donor if possible)

2) Titer is the tie breaker when applicable.
3) If no iSNV were found in a donor sample
}
\examples{


}
