% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_munging.R
\name{get_double}
\alias{get_double}
\title{Get all ENROLLID that have two SPECID in a season with same strain..}
\usage{
get_double(df, samples = 2)
}
\arguments{
\item{df}{A data frame with at least ENROLLID, SPECID, season, and pcr_result. Each row is an isolate.}

\item{samples}{The number of samples required to keep ENROLLID default = 2.}
}
\value{
the a data frame with only ENROLLID with multiple isolates - each row is an isolate
}
\description{
This is useful to get the samples we have intrahost data for.
}
\examples{
# There  one shared iSNV here and one fixed difference the rest are the same.
print(small_meta)

get_double(small_meta)

}
