% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{Get and compare the frequencies of mutations in two samples.}
\usage{
get_freqs(pairs, snv)
}
\arguments{
\item{pairs}{A vector of SPECID. These strings must be found in the SPECID column of the iSNV data frame}

\item{snv}{A data frame of isnv calls from deepSNV with a SPECID column}
}
\value{
a data frame with a row for each allele that differs between the two samples. Each loci that differs
should have two rows (one for each allele). If there are no iSNV or none that differ then an empty data frame
is returned.
}
\description{
This function compares the frequencies of alleles found in two samples. It relies
on equal_compare to make fair comaparisons at each loci.
}
\examples{
# There  one shared iSNV here and one fixed difference the rest are the same.
print(small_isnv)
get_freqs(c("HS1595","HS1563"),small_isnv)
}
\seealso{
\link{equal_compare}
}
