% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_munging.R
\name{finding_valid}
\alias{finding_valid}
\alias{getting_tp}
\title{Validating transmission pairs}
\usage{
finding_valid(tp)

getting_tp(meta_one)
}
\arguments{
\item{tp}{A data frame containing one row for each possible transmission pair}

\item{meta_one}{A data frame containing one row for each ENROLLID in the data set.}
}
\value{
A data frame with one row for each possible transmission pair

A data frame with one row for each possible transmission pair
}
\description{
Here we take a data frame where each row is a putative transmission pair
and determine which ones are valid. Valid pairs may only have the same
onset date if they are the index cases of the house. Such cases get a double
tag as well. Otherwise we only except pairs that do not jump a middle case,
and there can not be 2 possible donors (same onset date) for a recipient.
That is thrown out.

Here we take a data frame and look for all cases where two individuals are
sick within a week of each other. This does not compute wether or not the
pairs match our criteria yet
}
\section{Functions}{
\itemize{
\item \code{finding_valid}: 
}}

\examples{
one_meta<-only_one(small_meta)
getting_tp(one_meta)


}
