% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{like_found.beta}
\alias{like_found.beta}
\alias{like_lost.beta.uncert}
\alias{L.Nb.beta}
\title{Betabinomial Likelihood functions}
\usage{
like_found.beta(v_r, v_d, Nb, gc_ul, threshold, acc = accuracy_stringent)

like_lost.beta.uncert(v_r, v_d, Nb, gc_ul, threshold,
  acc = accuracy_stringent)

L.Nb.beta(v_r, v_d, Nb, gc_ul, threshold, acc = accuracy_stringent)
}
\arguments{
\item{v_r}{allele frequency in the recipient}

\item{v_d}{allele frequency in the donor}

\item{Nb}{Bottleneck size (may be a vector)}

\item{gc_ul}{titer in recipeint sample}

\item{threshold}{detection level threshold}

\item{acc}{a data frame with accrucacy data must contain, columns
freq,sensitivity, and gc_ul}
}
\value{
The likelihood of observing the data given the bottleneck size.
}
\description{
What is the probability of observing the data given a bottleneck
size under the beta binomial model. This is a wrapper around the two likelihood
functions
}
\section{Functions}{
\itemize{
\item \code{like_found.beta}: likelihood the variant is  found in recipient

\item \code{like_lost.beta.uncert}: likelihood the variant is not found in recipient
}}

