% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_munging.R
\name{longform_pairs}
\alias{longform_pairs}
\title{Convert short paired data to long paired data.}
\usage{
longform_pairs(df, columns = c("ENROLLID", "onset", "gc_ul", "sequenced"))
}
\arguments{
\item{df}{A data frame with obseravtions to be split}

\item{columns}{columns to be added. They must be equal to names that already exist
in every way except lacking either a 1 or a 2 at the end.}
}
\value{
the a data frame with a row for each ENROLLID
}
\description{
This function takes in a data frame containing short paired data with one
row / pair and outputs one row / ENROLLID in the pair. If the data does not
contain a column named 'pair_id' one will be added with a unqiue number for
each pair. This is currently groups by pair_id
}
\examples{
one_meta<-only_one(small_meta)
tp<-getting_tp(one_meta)
longform_pairs(tp)
}
