% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{math_fit}
\alias{math_fit}
\title{The probability of lambda for a loci}
\usage{
math_fit(data, l, Nb_max, model, threshold, acc)
}
\arguments{
\item{data}{a data frame of one donor polymorphic loci. It must have
columns chr,pos,freq1, and found}

\item{l}{a vector of lambda values to test}

\item{Nb_max}{The maximum bottleneck size to test}

\item{model}{The model we are using must be either "PA" or "BetaBin"}

\item{threshold}{limit of variant calling detection}

\item{acc}{a data frame with accuracy metrics}
}
\value{
a tibble with columns for each lambda and the probability of
observing the data for that site.
}
\description{
The probability of observing the data given lambda and the presence absence
model.In this fit we take the minority frequency to be  correct
}
\examples{
get_freqs(c("HS1595","HS1563"),small_isnv)->small_dups
polish_freq(small_dups,freq1,0.02)->x
x$found=c(TRUE,TRUE)
math_fit(x,1:2,100,"PA")
x$found=c(FALSE,TRUE)
math_fit(x,1:2,100,"PA")

}
