% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{model_summary}
\alias{model_summary}
\title{Summarize model likelihoods}
\usage{
model_summary(data)
}
\arguments{
\item{data}{data frame or tibble of model fit data.}
}
\value{
a tibble with columns
lambda : best fit
lower_95 : lower 95% bound (lambda)
upper_95 : upper 95% bound (lambda)
Nb : mean Nb of a zero truncated poisson given the lambda
}
\description{
This function summarizes the likelihood from the model_fit functions.
returning the max lambda and 95% confidence interval. The confidence interval
assumes a smooth curve.
}
\examples{
small_trans \%>\% pa_fit(1:2,100) \%>\% model_summary()
}
