% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_snv.R
\name{monomorphic}
\alias{monomorphic}
\title{Set monomorphic sites to freq =1}
\usage{
monomorphic(df, ...)
}
\arguments{
\item{df}{variant data frame}

\item{...}{columns to group by. There's no current need for them to be anything but
SPECID,season,chr,pos,pcr_result}
}
\value{
the input data frame but with monomorphic frequencies = 1.
}
\description{
In our analysis we calculate frequencies prior to
removing false positives. Sometimes this means we have a monomorphic site
with the allele present at <100%. I have checked all such sites between 50-90% and
in each case this holds true. That work is in 2017_11_13.Rmd- 2017_11_16.Rmd in the
notebook directory of the HIVE repo. It is also true we will miss some true variants
in the <10% range. Such is life. This script sets all sites that are monomorphic to
frequenies of 1.
}
\examples{
wacky_isnv<-small_isnv
wacky_isnv$freq.var[wacky_isnv$mutation=="PB2_G1602A"]<-0.8
monomorphic(wacky_isnv,SPECID,season,chr,pos,pcr_result)

}
