% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_model.R
\name{mut_model_LL}
\alias{mut_model_LL}
\title{Mutation rate and Ne from spectrum}
\usage{
mut_model_LL(data, mu, Ne, gen_time = 6, neg = FALSE,
  acc = accuracy_stringent)
}
\arguments{
\item{data}{The data frame containing iSNV calls. It must have the following columns:
freq.var,DPI-days post infection, and only contain 1 allele/ site. Sites not present will be assumed
to not have a variant present.}

\item{mu}{The mutation rate per nucleotide per cellular infectious cycle}

\item{Ne}{The effective population size}

\item{gen_time}{The generation time in hours. Default to 6}

\item{neg}{Boolean return the negative log likelihood not. Default =FALSE}

\item{acc}{An accuracy data frame}
}
\value{
The log likihood for the data given the parameters and the model
}
\description{
This function determines the Ne and mutation rate assuming all infections begin as a clone
and then mutations are built upon this base. The probability of a site going from 0 to
x given mu, Ne, and t generations given in Rouzine 2001 JVI as
}
\details{
\deqn{g(f,t) = \frac{2 \mu N_e}{f} e^{- \frac{2N_ef}{t}}}

This model makes some assumptions.
1) Every infection began with every potential loci with a mutation at frequency 0 - we only model minor alleles
2) The population develops under neutral processes
3) Because we intrate numerically we assume anything present at less than 0.001 (0.1%)
is essentially at 0.
4) There are 13133 sites that can contain mutations - the number of nucleotides in our concatenated 2014-2015 sample
}
