% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutational_model.R
\name{not_detected}
\alias{not_detected}
\title{Probability of not detecting a polymorphism
This is the probability the site is less than 0.001 (assumed 0 in this model) or
not found by our sequencing method.}
\usage{
not_detected(mu, Ne, t, gc_ul, acc = accuracy_stringent)
}
\arguments{
\item{mu}{The mutation rate per nucleotide per cellular infectious cycle}

\item{Ne}{The effective population size}

\item{t}{The number of generations}

\item{gc_ul}{The sample titer}

\item{acc}{dataframe holding the sensitivity}
}
\value{
The probability for the data given the parameters and the model
}
\description{
Probability of not detecting a polymorphism
This is the probability the site is less than 0.001 (assumed 0 in this model) or
not found by our sequencing method.
}
