% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_munging.R
\name{only_one}
\alias{only_one}
\title{Picking one sample per person}
\usage{
only_one(meta_df)
}
\arguments{
\item{meta_df}{A data frame containing one row for each SPECID in the data set.}
}
\value{
A data frame with one row for each ENROLLID in the data frame.
}
\description{
In the analysis it is sometimes useful to just have one sample per ENROLLID.
This is particularily true when we are simply interest in data specific to
the indivudal and not the sample. This function reduces a meta file that
may contain many SPECID/ENROLLID season pcr_result and reduces it so that just one SPECID
exists for each ENROLLID. We choose the best SPECID defined by
 the sequenced sample
 the sample that qualified for snv calling
 the sample with highest titer.
 If no samples were seqeunced we choose randomly.
}
\examples{
print(small_meta)
only_one(small_meta)


}
