% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{p_all}
\alias{p_all}
\title{Probability of drawing an allele n times in n draws}
\usage{
p_all(p, n)
}
\arguments{
\item{p}{The probability (frequency of allele)}

\item{n}{The number of draws. or a vector of draw values}
}
\value{
The probability of only drawing an allele at the given frequency.
Or a vector of such probabilities
}
\description{
Probability of drawing an allele n times in n draws. This is a helper
function used in transmission modeling.
}
