% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{pa_fit}
\alias{pa_fit}
\title{Fit presence absence model}
\usage{
pa_fit(data, l, Nb_max, ...)
}
\arguments{
\item{data}{a data frame or tibble it will be split by chr pos and pair_id}

\item{l}{a vector of lambda values to test}

\item{Nb_max}{The maximum bottleneck size to test}

\item{...}{other columns to group by in final output.}
}
\value{
a tibble with columns pair_id,lambda,LL (log likelihood), and pair_id if desired.
}
\description{
This is wrapper around math_fit that fits the presence absence model to
each pair present in the data set.
}
