% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{pa_sim}
\alias{pa_sim}
\title{Simulate transmission Presence absence model}
\usage{
pa_sim(data, lambda)
}
\arguments{
\item{data}{a data frame with with chr,pos,freq1, and  pair_id columns}

\item{lambda}{The lambda of the zero truncated Poisson}
}
\value{
data frame the same as data but with a simulated found column
}
\description{
Simulate the transmission of alleles using the
presence absence model. Essentially this takes
the frequency of the variant allele, and bottleneck
size and simulates the found column. It also selects a bottleneck size
based on lambda and a zero truncated Poisson. If simulatin the whole data
set each pair should be run separately as to not use the same bottleneck
}
\examples{
pa_sim(small_trans,1.3)
}
