% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{pa_sim_helper}
\alias{pa_sim_helper}
\title{PA sim helper}
\usage{
pa_sim_helper(data, Nb)
}
\arguments{
\item{data}{a data frame with with freq1 and pair_id columns and 2 rows}

\item{Nb}{the bottleneck size}
}
\value{
data frame the same as data but with a simulated found column
}
\description{
Helper function to simulate presence absence data. This takes in one loci.
the data frame should have 2 rows (one for each allele.)
Simulate the transmission of alleles using the
presence absence model. Essentially this takes
the frequency of the variant allele, and bottleneck
size and simulates the found column.
}
\examples{
pa_sim_helper(small_trans[1:2,],3)
}
