% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_snv.R
\name{read_rbind}
\alias{read_rbind}
\title{Read in a list of files}
\usage{
read_rbind(files, n = NULL, cols = NULL)
}
\arguments{
\item{files}{A list of files to read in and combine}

\item{n}{the optional entry to append as the run column. It is passed to split_path. It is the
nth entry in the path (sparating at '/')}

\item{cols}{An optional list of column types to pass to read_csv}
}
\description{
This function reads in a list of files and combines them into one tibble.
}
\details{
There are options to append a run column using a section of the path from
each file. When using we should speficy the Id column is a character as some
can be interpretted as either a character or a intereger. But others are characters.
}
\examples{
\dontrun{
read_rbind(c("../../Analysis/HIVE/data/processed/cali09/all.coverage.csv",
"../../Analysis/HIVE/data/processed/cali09_2/all.coverage.csv"),
n=7,
cols=list(Id=readr::col_character()))
}

}
