% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{sample_trans}
\alias{sample_trans}
\title{Randomly sample a transmission data.frame without any
restriction on donor SPECID}
\usage{
sample_trans(data, runs, pairs, test = F)
}
\arguments{
\item{data}{data frame on which to sample. contains freq1 and freq2}

\item{runs}{The number of times to sample}

\item{pairs}{The number of pairs to include in each run.}

\item{test}{A boolean switch. If activatve returns a list of 2 tibbles. The first is
the last sample_data tibble the second is the normal output}
}
\value{
a data frame with columns freq1, trial, and probability of transmission.
}
\description{
Each run we will ramdomly choose 'pairs' of the possible transmission pairs
from data  smoothing every 2%.
}
\examples{

sample_trans(small_community.comp,2,3)
}
