% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_munging.R
\name{short_pairs}
\alias{short_pairs}
\title{Convert long paired data to short paired data.}
\usage{
short_pairs(df, columns, ...)
}
\arguments{
\item{df}{A data frame with obseravtions to be combined}

\item{columns}{columns to be used in combination. They must exist and new columns
will be made appending 1 and 2 to these names}

\item{...}{unquoted grouping columns}
}
\value{
the a data frame with a row for each pair
}
\description{
This function takes in a data frame containing long paired data with one
row / sample in pair and outputs one row / pair. It is on the user to supply
grouping columns such that only 2 samples will match the groups. (we don't
handled more than 2 cases currently). If there are any columns that contain
data not shared by the sample pair they must be included or removed prior to using
the function
}
\examples{
get_double(small_meta)->small_doub
small_doub<-dplyr::select(small_doub,ENROLLID,pcr_result,HOUSE_ID,SPECID,
onset,collect,vaccination_status,DPI,season,gc_ul,sequenced,home_collected,snv_qualified)
columns = c("SPECID","onset","collect","vaccination_status",
           "DPI","gc_ul","sequenced","home_collected","snv_qualified")
short_pairs(small_doub,columns,ENROLLID,HOUSE_ID,pcr_result,season)
}
