% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{simulations}
\alias{simulations}
\title{Simulate probability of transmission  based on frequency}
\usage{
simulations(data, runs, lambda, FUN)
}
\arguments{
\item{data}{data frame of transmission data}

\item{runs}{how many simulations to run}

\item{lambda}{What is the lamda of the zero truncated Poisson to use}

\item{FUN}{what function to use to simulate the found column}
}
\value{

}
\description{
Simulate probability of transmission  based on frequency
}
