% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_community.comp}
\alias{small_community.comp}
\title{iSNV comparison between samples}
\format{a data frame with 32 rows and 13 variables
\describe{
   \item{chr}{Chromosome - genomic segment}
   \item{pos}{position on genomic segment}
   \item{SPECID1}{Specimen id for first sample}
   \item{SPECID2}{Specimen id for second sample}
   \item{mutation}{mutation as chr_refPosVar}
   \item{ref}{nucleotide in the plasmid control}
   \item{var}{nucleotide in the sample}
   \item{season}{The season of sampling}
   \item{pcr_result}{qpcr result for the samples-Strain}
   \item{freq1}{The frequency of the variant in the first sample}
   \item{freq2}{The frequency of the vairant in the second sample}
   \item{pair_id}{Id for the sample pairing}
   \item{found}{Is the variant found in the second sample}
}}
\source{
data-raw/small_community.comp.R
}
\usage{
small_community.comp
}
\description{
A silly example of what isnv comparisons between samples looks like. Note : this data
began as real and then was expanded to inlcude cases that are useful in testing code chunks.
It does not represent reality.
}
\keyword{datasets}
