% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_isnv}
\alias{small_isnv}
\title{iSNV present in samples}
\format{a data frame with 13 rows and 11 variables
#' \describe{
   \item{HOUSE_ID}{Id of the house of the individual}
   \item{ENROLLID}{Id of individual enrolled in cohort}
   \item{SPECID}{Specimen Id}
   \item{chr}{Chromosome - genomic segment}
   \item{pos}{position on genomic segment}
   \item{mutation}{mutation as chr_refPosVar}
   \item{ref}{nucleotide in the plasmid control}
   \item{freq.var}{Frequency of the variant}
   \item{var}{nucleotide in the sample}
   \item{season}{The season of sampling}
   \item{pcr_result}{qpcr result for the samples-Strain}
}}
\source{
data-raw/small_isnv.R
}
\usage{
small_isnv
}
\description{
A small example of the data frames used to hold isnv calls.
}
\keyword{datasets}
