% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_snv.R
\name{split_path}
\alias{split_path}
\title{Split a file path at /}
\usage{
split_path(x, n)
}
\arguments{
\item{x}{The string to be split.}

\item{n}{The entry to be return}
}
\description{
This function splits a string at '/' and returns
a one of the entries. You can not return muplitle entries at this point.
}
\details{
This is a hepler function in the read_rbind function.
}
\examples{
split_path("foo/bar/foo.csv",2)
}
