% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variants}
\alias{variants}
\title{Variant data}
\format{a data frame with 24 rows and 33 variables.
\describe{
   \item{Id}{The Id of the indexed sample for sequencing - can be multiple Id/ SPECID}
   \item{MapQ}{The average MapQ of reads with the variant}
   \item{Phred}{The average Phred of the variant base}
   \item{Read_pos}{The average position of the variant base on the read}
   \item{chr}{Chromosome - genomic segment}
   \item{cov.ctrl.bw}{The coverage of the plasmid control at the position in the reverse read}
   \item{cov.ctrl.fw}{The coverage of the plasmid control at the position in the forward read}
   \item{cov.tst.bw}{The coverage of the sample at the position in the reverse read}
   \item{cov.ctrl.fw}{The coverage of the sample at the position in the forward read}
   \item{freq.var}{Frequency of the variant}
   \item{mutation}{mutation as chr_refPosVar}
   \item{n.ctrl.bw}{The number of variant bases in the plasmid control in the reverse read}
   \item{n.ctrl.fw}{The number of variant bases in the plasmid control  in the forwad read}
   \item{n.tst.bw}{The number of variant bases in the sample in the reverse read}
   \item{n.ctrl.fw}{The number of variant bases in sample in the forward read}
   \item{p.val}{DeepSNV p.val}
   \item{pos}{position on genomic segment}
   \item{raw.p.val}{Uncorrected deepSNV p.val}
   \item{ref}{nucleotide in the plasmid control}
   \item{sigma2.freq.var}{Estimated variance of frequency measurement}
   \item{var}{nucleotide in the sample}
   \item{run}{The sequencing run the sample was on}
   \item{OR}{A list of the open reading frames overlapping this position}
   \item{coding_pos}{A list of the positions relative to the start of the in the ORs}
   \item{Ref_AA}{Reference amino acid - either sample or plasmid consensus depending on pipeline variables}
   \item{AA_pos}{Codon of the ORs - a list}
   \item{Var_AA}{Variant amino acid}
   \item{Class}{List of variant classifications in each relative OR}
   \item{LAURING_ID}{Sample ID used in processing - akin to SPECID}
   \item{dup}{Sequencing sample duplicate - if applicable}
   \item{coverage}{Total coverage at this position}
   \item{concat.pos}{Position in concatenated genome -
   segments in the order used in alignment fasta file}
   \item{gc_ul}{genome copy/ul of sample}
   }
   @source data-raw/variants.R}
\usage{
variants
}
\description{
And example of the iSNV data that comes out of the variant_calling_pipeline.
}
\keyword{datasets}
