% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{com_sample_trans}
\alias{com_sample_trans}
\title{Randomly sample a transmission data.frame based on a list
of donors}
\usage{
com_sample_trans(data, runs, SPECIDs, test = F)
}
\arguments{
\item{data}{data frame on which to sample. contains freq1 and freq2}

\item{runs}{The number of times to sample}

\item{SPECIDs}{A list of SPECIDs that can be used}
}
\value{
a data frame with columns freq1, trial, and probability of transmission.
}
\description{
Each run we will ramdomly choose one of the possible transmission pairs
from data with a donor ENROLLID provided in the list and evalute the probability
of transmission. Smoothing every 2%.
}
\examples{
print(small_comunity.comp)
com_sample_trans(small_community.comp,2,c("HS1595","MH0000"))
}
