\name{HiveR-package}
\Rdversion{1.1}
\alias{HiveR-package}
\alias{HiveR}
\docType{package}

\title{
2D and 3D Hive Plots for R
}

\description{
HiveR is an R package for creating and plotting 2D and 3D hive plots. Hive plots are a unique method of displaying networks of many types in which node properties are mapped to axes using meaningful properties rather than being arbitrarily positioned.  The hive plot concept was invented by Martin Kryzwinski at the Genome Science Center (www.hiveplot.com/).  Keywords: networks, food webs, linnet.
}

\details{
\tabular{ll}{
Package: \tab HiveR\cr
Type: \tab Package\cr
Version: \tab 0.2-10\cr
Date: \tab 2013-03-29\cr
License: \tab GPL-3\cr
}
}

\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{http://academic.depauw.edu/~hanson/HiveR/HiveR.html}
}
\keyword{ package }