% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveR-package.R
\docType{package}
\name{HiveR-package}
\alias{HiveR-package}
\alias{HiveR}
\title{2D and 3D Hive Plots for R}
\description{
Creates and plots 2D and 3D hive plots. Hive plots are a unique method of
displaying networks of many types in which node properties are mapped to
axes using meaningful properties rather than being arbitrarily positioned.
The hive plot concept was invented by Martin Krzywinski at the Genome
Science Center (www.hiveplot.net/).  Keywords: networks, food webs, linnet,
systems biology, bioinformatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bryanhanson/HiveR}
  \item Report bugs at \url{https://github.com/bryanhanson/HiveR/issues}
}

}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA
}
\keyword{package}
