% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Is this a whole number}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number}

\item{tol}{the tolerance used to compare x and round(x)}
}
\value{
Logical is the number an integer
}
\description{
Function that tets whether input is a whole number and
returns T/F. This is used creating the pmf of discrete
distributions. This is taken from the example in as.integer
documentation included in base.
}
