% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_models.R
\name{math_fit}
\alias{math_fit}
\title{The probability of lambda for a loci}
\usage{
math_fit(data, l = seq(0.01, 10, 0.01), Nb_max = 100)
}
\arguments{
\item{data}{a data frame of one donor polymorphic loci. It must have
columns chr,pos,freq1, and found}

\item{l}{a vector of lambda values to test}

\item{Nb_max}{The maximum bottleneck size to test}
}
\value{
a tibble with columns for each lambda and the probability of
observing the data for that site.
}
\description{
The probability of observing the data given lambda and the presence absence
model.In this fit we take the minority frequency to be  correct
}
\examples{
polish_freq(small_dups,freq1,0.02)->x
x$found=c(T,T)
math_fit(x,1:2)

x$found=c(F,T)
math_fit(x,1:2)

}
